+++ src/lib/types.ts
@@ -21,19 +21,17 @@ export interface ToastFunctionOptions {
 	duration?: number;
 	/** A custom component to be rendered. */
 	component?: ToastCustomComponent;
+	/** Never remove the toast. */
+	infinite?: boolean;
 }
 
-export interface ToastComponent {
+export interface ToastComponent extends Required<Omit<ToastFunctionOptions, 'component'>> {
 	/** The unique ID of the toast. */
 	id: number;
 	/** The attention level of the toast. */
 	type: ToastType;
 	/** The message to display to the end user. */
 	message: string;
-	/** Allow the toast to be closable. */
-	closable: boolean;
-	/** The duration of the toast in milliseconds */
-	duration: number;
 }
 
 export interface ToastComponentWithCustom extends ToastComponent {